Correctly detect the application type
Pick up the application channel correctly so that the deeplink URL is built correctly
