+++ crates/but/Cargo.toml
@@ -38,6 +38,7 @@ but-serde.workspace = true
 but-github.workspace = true
 but-secret.workspace = true
 but-oxidize.workspace = true
+but-path.workspace = true
 
 gitbutler-project.workspace = true
 gitbutler-commit.workspace = true
@@ -76,7 +77,6 @@ ratatui = "0.29.0"
 crossterm = "0.29.0"
 atty = "0.2.0"
 cli-prompts = "0.1.0"
-open.workspace = true
 
 [dev-dependencies]
 but-core = { workspace = true, features = ["testing"] }
